chore: changes the contents of the disciplines.test.ts and teachers.test.ts files to the tests.test.ts file
